# Audio: Use generic saturation logic for improved efficiency