Introduce a arithmetic bitwise saturation function operations across different integer sizes (8, 16, 24, 32 bits) in the 
processing block. 

- Replaced `if-else` checks with bitwise masks to handle overflow and
  underflow for int32, int24, int16, and int8 saturation functions.

- For each bit width, created masks by shifting the results of comparisons
  (x - `MIN_VAL`) and (`MAX_VAL` - x) to the respective overflow/underflow
  detection bits.

- Mask >> `31` or >> `63` results in either `0` (no overflow/underflow) or `-1`
  (overflow/underflow occurred).

- Applied bitwise operations to conditionally replace original values
  with boundary values (`MIN` or `MAX`) based on the computed masks.

- This approach avoids branching, improves efficiency, and ensures
  accurate saturation within specified bit-width limits.
  
**_HiFi3 Inline Summary_**
sat_int32: Decreased from 17 to 6-9 cycles; Bit true: Mixed; Gain: ~47.06% to 64.71%
sat_int24: Decreased from 7 to 5 cycles; Bit true: Yes; Gain: ~28.57%
sat_int16: Decreased from 13 to 7 cycles; Bit true: Yes; Gain: ~46.15%
sat_int8: Decreased from 3 to 1 cycle; Bit true: Yes; Gain: ~66.67%
 
**_HiFi4 Inline Summary_**
sat_int32: Decreased from 11 to 6-8 cycles; Bit true: Mixed; Gain: ~27.27% to 45.45%
sat_int24: No change (5 cycles); Bit true: Yes
sat_int16: Decreased from 9 to 7 cycles; Bit true: Yes; Gain: ~22.22%
sat_int8: Decreased from 3 to 1 cycle; Bit true: Yes; Gain: ~66.67%
 
_**HiFi5 Inline Summary**_
sat_int32: Decreased from 11 to 6-8 cycles; Bit true: Mixed; Gain: ~27.27% to 45.45%
sat_int24: No change (5 cycles); Bit true: Yes
sat_int16: Decreased from 9 to 7 cycles; Bit true: Yes; Gain: ~22.22%
sat_int8: Decreased from 3 to 1 cycle; Bit true: Yes; Gain: ~66.67%
 
**_Summary_**: Inline functions showed significant performance gains across all architectures,
 especially in cycle reductions, while noinline functions maintained consistent cycle counts with bit true results.
 
 
![image](https://github.com/thesofproject/sof/assets/61411947/59ee8937-1872-481d-8724-63bdc8338d81)