@ShriramShastry 2024-07-05
> optimisation attempt seems to have failed

With latest modification,  the gap in performance is addressed i.e. Before and After the changes, there's no performance difference

_**ACE-3**_

`There's gain of 1 cycle count. i.e. sat_int32: decreased from 21 to 20 cycles`

![image](https://github.com/thesofproject/sof/assets/61411947/8eae61fb-353d-499f-97fb-bf842f3a8054)

_**Generic HiFi5**_ 

![image](https://github.com/thesofproject/sof/assets/61411947/cc5fa430-76d5-4ecd-b5a2-efc74c082623)


_**Generic HiFi4**_ 

![image](https://github.com/thesofproject/sof/assets/61411947/5d46976e-7436-44b9-860b-6ae67a33e9f3)


_**Generic HiFi3**_ 

![image](https://github.com/thesofproject/sof/assets/61411947/b0ba534a-9478-49ea-a34a-8f089a8b8a00)