@lgirdwood 2024-07-05
> There's gain of 1 cycle count. i.e. sat_int32: decreased from 21 to 20 cycles

@ShriramShastry  this is the sort of data we need for every change as its not really clear from all the table (there is too much data).  We just need to know the before and after in a simplified form like your comment above. 

> Here's a summary of the cycles before and after the modification :
```
sat_int32: No change (20 cycles both before and after)
sat_int24: Increased from 16 to 21 cycles
sat_int16: Increased from 13 to 21 cycles
sat_int8: Increased from 13 to 20 cycles
```
Ok, so this looks like a small gain for int32, but a bigger loss for 24, 16 and 8 and means the PR should not be merged until we can show no loss for any format. Do you have a plan to rework ?