@lyakh 2024-07-08
> I have reworked on sat-24,16 and 8 code and addressed performance loss !! With latest check-in there is no performance difference between `Before` and `After` my modification.

so, there's no need for this PR then. We don't make changes just because they don't make things worse. We only make changes if they make something better.