@ShriramShastry 2024-07-08
> > I have reworked on sat-24,16 and 8 code and addressed performance loss !! With latest check-in there is no performance difference between `Before` and `After` my modification.
> 
> so, there's no need for this PR then. We don't make changes just because they don't make things worse. We only make changes if they make something better.


Thank you for your feedback on the recent changes. I would like to address your concerns regarding performance.

**_Summary of Changes_**
The performance evaluation indicates that there is `no observable performance loss` due to the recent changes. Below is a summary of results from the performance evaluation:

![image](https://github.com/thesofproject/sof/assets/61411947/c087b965-b2e6-4ea7-8044-cb9f6ea49052)

**Benefits of the Optimization**
The aim of this optimization was not solely to enhance performance but also to enhance code reliability and maintainability. Here are the key benefits of the bitwise approach:

**Consistent Performance:**

There is no performance degradation; the cycle counts before and after the optimization remain the same across all tested values.

**Branch Prediction:**

The use of bitwise operations eliminates branches, which can reduce the risk of branch mispredictions and enhance predictability in pipeline performance.

**Code Clarity and Maintenance:**

The bitwise approach makes the code more compact and potentially less error-prone, improving readability and future maintainability.

**Future-proofing:**

Branch-free code is often favored in modern compiler and processor optimizations, which might yield further advantages on future architectures.

**Conclusion**
To summarize, the optimization has no negative impact on performance, maintains the same cycle counts as the original implementation, and provides additional long-term benefits in terms of code quality and predictability. Therefore, this change does not degrade performance and is a worthwhile improvement.