@marc-hb 2024-07-08
> To summarize, the optimization has no negative impact on performance, maintains the same cycle counts as the original implementation

Then you need to update the commit title and commit message. As they stand:

> This approach avoids branching, **improves efficiency**, and ensures
>  accurate saturation within specified bit-width limits.

Was there some accuracy issue that this is fixing?

This seems missing from the commit message too:
> which might yield further advantages on future architectures.

If this just a code clean-up  then say "more readable code" and don't mention performance at all. If it's more than a clean-up then you need some data or reference.