@singalsu 2024-07-18
I did a performance comparison with git main vs main + #9170 with TGL build with gcc and xcc. The gcc build currently can't run enabled DRC (reason unknown, a regression or just too much load) so I used different configurations for hw:0,0 playback on sof-hda-generic-4ch.tplg

GCC (gain 1.1 44, gain 2.1 44, EQIIR eq_iir_flat.txt, EQFIR eq_fir_flat.txt, DRC passthrough.txt  off)
- main 69.8 MCPS
- 9170 74.2 MCPS

XCC (gain 1.1 44, gain 2.1 44, EQIIR eq_iif_spk.txt, EQFIR eq_fir_spk.txt, DRC  threshold_-25_knee_15_ratio_10.txt on)
- main 96.8 MCPS
- 9170 96.8 MCPS

I'll see if I can repeat this with some remote MTL device. But this doesn't look for TGL.