@singalsu 2024-07-18
Similar result with MTL, no improvement with xcc build,  and about same 4 MCPS slower with gcc build of this PR.