@andyross 2024-07-18
So... I hate to say it but I'm pretty sure this change is just a big noop to the optimizer.  Here's some test code that extracts the two variants in individually compilable form, and a just-non-trivial-enough outer loop to let the compiler try to use the code in a natural setting.  And the generated code from xt-clang (RI-2022.10, at -Os) is essentially identical.  Both variants reduce to SALTU instructions.  There is no branching in the current code nor masking in the new one; the compiler is smart enough to recognize what's happening.

```c
#define INT_MAX 0x7fffffff
#define INT_MIN 0x80000000

static inline int sat32_cmp(long long x)
{
	return x > INT_MAX ? INT_MAX : (x < INT_MIN ? INT_MIN : x);
}

static inline int sat32_mask(long long x)
{
	long long mask_overflow = (INT_MAX - x) >> 63;
	long long mask_underflow = (x - INT_MIN) >> 63;

	x = (x & ~mask_overflow) | (INT_MAX & mask_overflow);
	x = (x & ~mask_underflow) | (INT_MIN & mask_underflow);
	return (int) x;
}

void sat_array_add_cmp(int *dst, int *a, int *b, int n)
{
	for (int i = 0; i < n; i++) {
		dst[i] = sat32_cmp(a[i] + (long long) b[i]);
	}
}

void sat_array_add_mask(int *dst, int *a, int *b, int n)
{
	for (int i = 0; i < n; i++) {
		dst[i] = sat32_mask(a[i] + (long long) b[i]);
	}
}
```